Update markdown config
- Update eleventy.config.js
- Add markdown.js
- Update filters.js
